backend/requirements.txt
Flask
flask-cors
pyjwt
